requirements.txt
langchain==0.0.304
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.27.1
faiss-cpu==1.7.4
altair==5.1.1
openai==0.28.1
tiktoken==0.5.1